mssql/requirements.txt
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_0ek9yztvu3/croot/certifi_1665076692562/work/certifi
numpy==1.23.4
pandas==1.5.1
pymssql==2.2.5
pyodbc==4.0.34
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
